+++ eslint.config.mjs
@@ -1,36 +1,43 @@
-import { defineConfig, globalIgnores } from "eslint/config";
-import tsParser from "@typescript-eslint/parser";
-import path from "node:path";
-import {fileURLToPath} from 'node:url'
-import js from "@eslint/js";
-import { FlatCompat } from "@eslint/eslintrc";
-import globals from 'globals'
+import antfu from '@antfu/eslint-config'
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const compat = new FlatCompat({
-  baseDirectory: __dirname,
-  recommendedConfig: js.configs.recommended,
-  allConfig: js.configs.all,
-});
-
-export default defineConfig([globalIgnores(["**/dist/"]), {
-  extends: compat.extends(
-    "eslint:recommended",
-    "plugin:@typescript-eslint/eslint-recommended",
-    "plugin:@typescript-eslint/recommended",
-  ),
-
-  languageOptions: {
-      globals: {
-          ...globals.browser,
-          ...globals.commonjs,
-      },
-
-      parser: tsParser,
-      ecmaVersion: "latest",
-      sourceType: "script",
+export default antfu(
+  {
+    ignores: ['dist', 'info', 'ui/.angular'],
+    rules: {
+      'curly': ['error', 'all'],
+      'no-undef': 'error',
+      'perfectionist/sort-exports': 'error',
+      'perfectionist/sort-imports': [
+        'error',
+        {
+          groups: [
+            ['type-builtin', 'type-external', 'type-internal'],
+            ['type-parent', 'type-sibling', 'type-index'],
+            'builtin',
+            'external',
+            'internal',
+            ['parent', 'sibling', 'index'],
+            'side-effect',
+            'unknown',
+          ],
+          internalPattern: ['^@/.*'],
+          order: 'asc',
+          type: 'natural',
+          newlinesBetween: 1,
+        },
+      ],
+      'perfectionist/sort-named-exports': 'error',
+      'perfectionist/sort-named-imports': 'error',
+      'style/brace-style': ['error', '1tbs'],
+      'style/quote-props': ['error', 'consistent-as-needed'],
+      'test/no-only-tests': 'error',
+      'ts/consistent-type-imports': 'off',
+      'unicorn/no-useless-spread': 'error',
+      'unused-imports/no-unused-vars': ['error', { caughtErrors: 'none' }],
+    },
+    typescript: true,
+    formatters: {
+      markdown: true,
+    },
   },
-
-  rules: {},
-}]);
+)