+++ src/index.test.ts
@@ -1,7 +1,10 @@
-import { EventEmitter } from 'events';
-import { HapClient } from './index';
+import { EventEmitter } from 'node:events'
 
-jest.mock('axios');
+import { afterEach, beforeEach, describe, expect, it, jest } from '@jest/globals'
+
+import { HapClient } from './index'
+
+jest.mock('axios')
 
 jest.mock('bonjour-service', () => {
   return jest.fn().mockImplementation(() => ({
@@ -10,23 +13,23 @@ jest.mock('bonjour-service', () => {
     find: jest.fn().mockReturnValue({
       start: jest.fn(),
       stop: jest.fn(),
-      on: jest.fn()
-    })
-  }));
-});
+      on: jest.fn(),
+    }),
+  }))
+})
 
-describe('HapClient', () => {
-  let hapClient;
+describe('hapClient', () => {
+  let hapClient
 
   beforeEach(() => {
-    hapClient = new HapClient({ pin: '123-45-678', config: {} });
-  });
+    hapClient = new HapClient({ pin: '123-45-678', config: {} })
+  })
 
-  test('should initialize correctly', () => {
-    expect(hapClient).toBeInstanceOf(EventEmitter);
-  });
+  it('should initialize correctly', () => {
+    expect(hapClient).toBeInstanceOf(EventEmitter)
+  })
 
   afterEach(() => {
-    hapClient.destroy();
-  });
-});
+    hapClient.destroy()
+  })
+})