+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `@homebridge/hap-client` will be documented in this file. This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
+## v3.2.1 (Unreleased)
+
+### Changed
+
+- bump `axios` from `v1.13.4` to `v1.13.5` (#30) (@dependabot)
+- update lint and code style + eslint dependencies
+
 ## v3.2.0 (2026-02-08)
 
 ### Changed